promote aiProvider into services, keeping its pure text helpers in lib
lib/aiProvider.js imported four services — providers, aiStatusEvents,
ollamaManager, mtplxServerManager — to resolve a provider and call it. That is
provider orchestration, so it moves up rather than having those dependencies
pushed down.

Two of its exports had no business moving with it: stripCodeFences and
parseLLMJSON are pure string work, and two lib modules (jsonExtract.js,
stageRunner.js) use them. Promoting the whole file would have inverted those in
turn. They move DOWN to the new lib/llmText.js instead, and services/aiProvider.js
re-exports both so the dozen services that have always imported them from there
keep working.

So one file splits in both directions: the provider calls go up, the text
helpers go down, and the lib modules that only ever wanted to unfence a string
no longer reach through provider orchestration to do it.

Refs #4901.
