+++ server/services/taste-questionnaire.test.js
@@ -19,7 +19,7 @@ vi.mock('./digital-twin.js', () => ({
 // Drive the provider boundary directly so each "Go deeper" outcome is reachable
 // without a live LLM. Re-imported per test because beforeEach resets the module
 // registry, which mints fresh spies for each generation.
-vi.mock('../lib/aiProvider.js', () => ({
+vi.mock('./aiProvider.js', () => ({
   resolveAPIProvider: vi.fn(),
   callProviderAISimple: vi.fn(),
 }));
@@ -140,7 +140,7 @@ describe('generatePersonalizedTasteQuestion outcomes', () => {
   });
 
   it('reports no-provider rather than blaming the user documents that do exist', async () => {
-    const { resolveAPIProvider } = await import('../lib/aiProvider.js');
+    const { resolveAPIProvider } = await import('./aiProvider.js');
     resolveAPIProvider.mockResolvedValue(null);
     const taste = await import('./taste-questionnaire.js');
     await seedIdentityContext(taste);
@@ -152,7 +152,7 @@ describe('generatePersonalizedTasteQuestion outcomes', () => {
   });
 
   it('throws AI_PROVIDER_ERROR instead of collapsing a provider failure into "nothing to ask"', async () => {
-    const { resolveAPIProvider, callProviderAISimple } = await import('../lib/aiProvider.js');
+    const { resolveAPIProvider, callProviderAISimple } = await import('./aiProvider.js');
     resolveAPIProvider.mockResolvedValue(PROVIDER);
     callProviderAISimple.mockResolvedValue({ error: 'Provider returned 401: invalid key' });
     const taste = await import('./taste-questionnaire.js');
@@ -173,7 +173,7 @@ describe('generatePersonalizedTasteQuestion outcomes', () => {
   });
 
   it('returns the question with reason null on success', async () => {
-    const { resolveAPIProvider, callProviderAISimple } = await import('../lib/aiProvider.js');
+    const { resolveAPIProvider, callProviderAISimple } = await import('./aiProvider.js');
     resolveAPIProvider.mockResolvedValue(PROVIDER);
     callProviderAISimple.mockResolvedValue({ text: '  Which film would you rewatch forever?  ' });
     const taste = await import('./taste-questionnaire.js');
