+++ server/services/universeStyleReference.js
@@ -7,7 +7,7 @@
  */
 
 import { randomUUID } from 'crypto';
-import { parseLLMJSON, resolveAPIProvider } from '../lib/aiProvider.js';
+import { parseLLMJSON, resolveAPIProvider } from './aiProvider.js';
 import { ServerError } from '../lib/errorHandler.js';
 import { assertProvider, assertVisionRunUsedImages, runPromptThroughProvider } from '../lib/promptRunner.js';
 import { trimTo } from '../lib/storyBible.js';
